refactor(executor): split mod.rs into focused modules
